requirements.txt
langchain
langchain-openai
langchain-community
langchain-core
langchain-experimental
langgraph
langsmith
faiss-cpu
fastapi
uvicorn
python-dotenv
pydantic
typing-extensions
tiktoken
python-multipart
beautifulsoup4
groq
requests
sse-starlette
langgraph-checkpoint-sqlite
